- api/fabric_itf.h:    - move uses of "context" from fabric-defined to user-defined    - add memory regsion key to register_memory return value, as      the API user will need it    - introduce an "IFabric" class to cover everything in libfabric above a resource domain  - net/fabric/src    - add compiler warnings, and guards around external code which fails those warnings    - provide seperate subclasses for client and server fabric connections    - move fabric_error function definitions out of line    - rename utility code from fabric_help to fabric_util    - fix negotiation of fabric connections (fi_connect, fi_accept)    - add test of client/server RDMA function

+++ src/components/net/fabric/src/fd_socket.cpp
@@ -14,9 +14,9 @@
    limitations under the License.
 */
 
-/* 
- * Authors: 
- * 
+/*
+ * Authors:
+ *
  */
 
 #include "fd_socket.h"
@@ -52,7 +52,7 @@ Fd_socket::~Fd_socket()
 Fd_socket::Fd_socket(Fd_socket &&o) noexcept
   : _fd(o._fd)
 {
-  o._fd = -1;  
+  o._fd = -1;
 }
 
 Fd_socket &Fd_socket::operator=(Fd_socket &&o) noexcept
@@ -87,9 +87,10 @@ void Fd_socket::send(const void *buf, std::size_t size) const
      system_fail(ECONNABORTED, "send");
   }
 }
+
 void Fd_socket::recv(void *buf, std::size_t size) const
 {
-  int r;
+  ssize_t r;
   do
   {
      r = ::read(_fd, buf, size);
